+++ apps/api/src/__tests__/services/import-entity-ref.service.integration.test.ts
@@ -9,11 +9,9 @@ import { brandId } from "@pluralscape/types";
 import { drizzle } from "drizzle-orm/pglite";
 import { afterAll, afterEach, beforeAll, describe, expect, it } from "vitest";
 
-import {
-  listImportEntityRefs,
-  lookupImportEntityRef,
-  recordImportEntityRef,
-} from "../../services/import-entity-ref.service.js";
+import { listImportEntityRefs } from "../../services/import-entity-ref/list.js";
+import { lookupImportEntityRef } from "../../services/import-entity-ref/lookup.js";
+import { recordImportEntityRef } from "../../services/import-entity-ref/record.js";
 import { asDb, assertApiError, makeAuth } from "../helpers/integration-setup.js";
 
 import type { AuthContext } from "../../lib/auth-context.js";
