+++ apps/api/src/trpc/routers/import-entity-ref.ts
@@ -2,12 +2,12 @@ import { IMPORT_ENTITY_TYPES, IMPORT_SOURCES } from "@pluralscape/types";
 import { z } from "zod/v4";
 
 import { MAX_PAGE_LIMIT } from "../../service.constants.js";
+import { listImportEntityRefs } from "../../services/import-entity-ref/list.js";
 import {
-  listImportEntityRefs,
   lookupImportEntityRef,
   lookupImportEntityRefBatch,
-  upsertImportEntityRefBatch,
-} from "../../services/import-entity-ref.service.js";
+} from "../../services/import-entity-ref/lookup.js";
+import { upsertImportEntityRefBatch } from "../../services/import-entity-ref/upsert-batch.js";
 import { createTRPCCategoryRateLimiter } from "../middlewares/rate-limit.js";
 import { systemProcedure } from "../middlewares/system.js";
 import { router } from "../trpc.js";
